fix(angular-base): adding event handlers for angular
