+++ packages/teleport-shared/src/node-handlers/node-to-html/utils.ts
@@ -81,8 +81,12 @@ export const handleEvent = (
     const methodName = `handle${dashCaseToUpperCamelCase(elementName)}${dashCaseToUpperCamelCase(
       eventKey
     )}`
+    const eventNameBiding = templateSyntax.eventNameBinding
+      ? templateSyntax.eventNameBinding(methodName)
+      : methodName
+
     methodsObject[methodName] = eventHandlerStatements
-    htmlUtils.addAttributeToNode(htmlNode, eventHandlerKey, methodName)
+    htmlUtils.addAttributeToNode(htmlNode, eventHandlerKey, eventNameBiding)
   }
 }
