+++ packages/teleport-shared/src/utils/ast-js-utils.ts
@@ -1,4 +1,5 @@
 import * as types from '@babel/types'
+import { UIDLEventHandlerStatement, UIDLPropDefinition } from '@teleporthq/teleport-types'
 /**
  * A tricky way to pass down custom configuration into
  * the objectToObjectExpression values, to allow for member expressions like
@@ -110,4 +111,73 @@ export const getTSAnnotationForType = (type: any, t = types) => {
     default:
       return t.tsUnknownKeyword()
   }
+}
+
+export const createMethodsObject = (
+  methods: Record<string, UIDLEventHandlerStatement[]>,
+  propDefinitions: Record<string, UIDLPropDefinition>,
+  flavor: 'angular' | 'vue',
+  t = types
+) => {
+  return Object.keys(methods).map((eventKey) => {
+    const astStatements = []
+    methods[eventKey].map((statement) => {
+      const astStatement =
+        statement.type === 'propCall'
+          ? createPropCallStatement(statement, propDefinitions)
+          : createStateChangeStatement(statement)
+
+      if (astStatement) {
+        astStatements.push(astStatement)
+      }
+    })
+    return flavor === 'angular'
+      ? t.classMethod('method', t.identifier(eventKey), [], t.blockStatement(astStatements))
+      : t.objectMethod('method', t.identifier(eventKey), [], t.blockStatement(astStatements))
+  })
+}
+
+const createStateChangeStatement = (statement: UIDLEventHandlerStatement, t = types) => {
+  const { modifies, newState } = statement
+
+  const rightOperand =
+    newState === '$toggle'
+      ? t.unaryExpression('!', t.memberExpression(t.identifier('this'), t.identifier(modifies)))
+      : convertValueToLiteral(newState)
+
+  return t.expressionStatement(
+    t.assignmentExpression(
+      '=',
+      t.memberExpression(t.identifier('this'), t.identifier(modifies)),
+      rightOperand
+    )
+  )
+}
+
+const createPropCallStatement = (
+  eventHandlerStatement: UIDLEventHandlerStatement,
+  propDefinitions: Record<string, UIDLPropDefinition>,
+  t = types
+) => {
+  const { calls: propFunctionKey, args = [] } = eventHandlerStatement
+
+  if (!propFunctionKey) {
+    console.warn(`No prop definition referenced under the "calls" field`)
+    return null
+  }
+
+  const propDefinition = propDefinitions[propFunctionKey]
+
+  if (!propDefinition) {
+    console.warn(`No prop definition was found for function "${propFunctionKey}"`)
+    return null
+  }
+
+  // In vue it's favorable to use $emit for a specific event than sending the function as a prop
+  return t.expressionStatement(
+    t.callExpression(t.identifier('this.$emit'), [
+      t.stringLiteral(propFunctionKey),
+      ...args.map((arg) => convertValueToLiteral(arg)),
+    ])
+  )
 }
